Deprecate GitHub Pull Request Builder plugin (#8)
The GitHub Pull Request Builder plugin should be deprecated so that
users do not mistakenly install it when integrating with GitHub.

More details are available in Jenkins infrastructure help desk ticket:

* https://github.com/jenkins-infra/helpdesk/issues/4608

The deprecation message is being added here because a new release of the
plugin would be required in order to move documentation to the plugin
GitHub repository.  We don't want a new release of the plugin just to
declare it deprecatd.

A similar message will be proposed for the README in the plugin repository.